previous level is not newly adjacent
This is a bit of an edge case. Not sure if this will break something
later on, but decided not to treat the previous level as newly adjacent
when entering another room.

I guess technically the room does become newly adjacent, but on the
other hand the adjacency is meant to convey the message 'level became
visible', not 'level became adjacent to the current level'.

I'll probably regret this later on.