support entrance offsets for level transitions
This allows creating the levels positioned at the origin and requires
worrying about the relative offsets only once transitions are
configured. In practice, this means that levels can be created
indenpendently without having to position the scenes at awkward
coordinates.

Got an excuse to do some funky scene file queries (reading nodes from a
scene without instantiating it in-editor), which turned out quite fun.
There is now a programmatically created enum (dropdown) property for
level transitions, which lists all nodes in 'LevelEntrance' group as
options. Then, there are a bunch of other magic properties for handling
the backing storage to the scene file and whatnot.

While at it, converted the scene preview to a programmatically created
property, too, as it is not really a value stored inside the node, but
rather a debug toggle.

There is quite a bit of magical boilerplate related to all this. I hope
to get around to clean up things a bit at some point.