mostly well-behaved scene transition editor
It's always pleasantly surprising how much Godot exposes of its
internals to the scripts. Turns out, one can quite easily query the
.tscn resource file without instantiating it, in order to determine
details about the nodes in a given scene. Also turns out things get a
bit trickier when one wants to perform write operations on the given
scene.

In order to get modifying the scene entrance via other scene's level
transition working, we need to be able to do arbitrary edits to a scene
file, while editing another one. To my surprise and shock, this wasn't
even as convoluted as I feared.

In practice, the steps to modify the Other Scene are:
 1. load the other scene PackedScene resource with the ResourceLoader
 2. instantiate the other scene (already done for the preview)
 3. make modifications to the instantiated copy (need to undo any
    preview-specific modifications first)
 4. re-pack the instantiated copy into a PackedScene
 5. save the PackedScene resource to a .tscn file via ResourceSaver
 6. overtake the resource path with the modified scene (overwrites the
    cache entry with the modified one)
 7. force editor to reload the scene resource (ensures open editor tabs
    are refreshed)
 8. (re-apply preview-specific modifications to the instantiated copy to
    keep the preview working)

The steps 1 and 2 are more-or less done already, as the preview also has
to load and instantiate the scene. This creates slight additional
complexity as e.g. scene root node process mode and position are adjusted
for the preview, but should remain unchanged when the modifications are
saved on disk. This requires tracking the original values of those
properties and restoring them temporarily when the scene is saved.

Most of the steps are clear and there was little to no issues
implementing those.

The tricky part was getting the resource loader caches and the editor to
behave nicely with the "external" modifications to the scene files.

The save/load happens eagerly and there is no caching of the resources
whatsoever , resulting in unnecessary roundtrips to ResourceLoader.
Also, as the save occurs eagerly immediately as the change occurs, the
editor freezes for a moment there, making modifying the other scene
offsets a bit janky.

But it works!