adjacent level preloading v0.1
This already works for the simple test scene, but I expect things to
break when there are more adjacent scenes. Scene unloading is not
properly implemented and/or is not working. The code is getting a bit
a spaghetti-y, so it is hard to tell. That said, there is a fair bit of
cleanup to do, too.