port intro sequence to use the new player rig
To get rid of a whole class of bugs and oddities (and surely, to
introduce a sleuth of new ones instead!), I decided to get rid of the
separate player sprite for the intro sequence. This streamlines the
transitions to/from cutscenes, as the cutscene player *is* the actual
player character.

This creates complications for previewing animations in-editor. The
player instance lives in the persistent scene during game
initialization, so it cannot directly be referenced from the cutscene
animation player(s).

I solved this by creating a EditorOnly-node, which automagically removes
its parent during _enter_tree() if running in-game. Then, I could create
a placeholder player to the intro scene and animate against that. A
EditorOnly component node yeets the placeholder player automagically,
leaving the animation track node paths dangling. Before the animation
player is activated, the real player is reparented to the same path
where the placeholder was, so the animation tracks now point to it. As
animation player caches are cooked only when it is first used, this
approach works as long as no-one tries to do anything with the animation
player before the player has been properly reparented.

Using this, alongside the new player rig thingamajig, it was quite
straightforward to animate the intro cutscene using the real player
instead of a "fake" sprite.

Intro script required minimal changes, but I did a bunch of refactoring
etc. improvements, to make the intro a bit smoother overall.

Next up camera controls, another dialogue system rewrite, and wisp
collision avoidance (possibly).